+++ src/compounds/legacy-product-table/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.legacy-product-table@0.2.0...@uswitch/trustyle.legacy-product-table@0.2.1) (2020-11-24)
+
+
+### Bug Fixes
+
+* add maxHeight to image and marginTop to first row ([9770cd8](https://github.com/uswitch/trustyle/commit/9770cd8))
+* change first-child to first-of-type ([5e6f975](https://github.com/uswitch/trustyle/commit/5e6f975))
+* change first-child to first-of-type ([61b55fd](https://github.com/uswitch/trustyle/commit/61b55fd))
+
+
+
+
+
+# [0.2.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.legacy-product-table@0.1.2...@uswitch/trustyle.legacy-product-table@0.2.0) (2020-11-24)
+
+
+### Bug Fixes
+
+* pass link to cta and clickable row ([66e1924](https://github.com/uswitch/trustyle/commit/66e1924))
+
+
+### Features
+
+* add clickable row wrapper ([3310069](https://github.com/uswitch/trustyle/commit/3310069))
+
+
+
+
+
 ## [0.1.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.legacy-product-table@0.1.1...@uswitch/trustyle.legacy-product-table@0.1.2) (2020-11-23)
 
 **Note:** Version bump only for package @uswitch/trustyle.legacy-product-table